Initial commit
As of right now, GitVersion is mostly a port of the Git utilities from
GradleUtils and SharedActions. This is to ensure that the versioning
system is consistent across all Forge projects.

While not directly credited as co-authors, GitVersion was built off of
the work done in GradleUtils and SharedActions by LexManos and prior
Forge contributors.