Valid tags give a default offset of 0
Previously, tags that were created for a subproject needed to include a
commit that modified a file within the path. This is no longer required
and as long as the tag is created for the project's tag prefix, it will
be counted as part of the project.

For example, for project hash-utils with tag prefix 'hash', we have
three commits:
1. An empty commit tagged 'hash-0.3'
2. A commit that modifies a file in the 'hash-utils' directory
3. A commit that does not modify any files in 'hash-utils'

Commits 1 and 2 would be counted, but 3 would not. The resulting
version number will be 0.3.1.

This is currently using an implementation detail regarding how JGit's
LogCommand returns its result. See eclipse-jgit/jgit#149