Preliminary support for tag suffixes
Instead of adding some more convoluted and additional code to Git
Version specifically to handle tags with suffixes, it instead now uses a
new approach when handling tag prefix.

tl;dr default filter `*-*` is now `[[:digit"]]**`.

Even since before Git Version, it was established through the use of
`Character#isDigit` that all tags should, in some way, include a valid
number to convey the version. Prior to Git Version 0.4, when checking
for tags without a tag prefix, the filter would simply check if the tag
didn't have any dashes in it. Since the tag prefix should only ever
remain a prefix, I've simplified this to instead check for a digit.
JGit's glob filtering uses the same digit detection as
`Character#isDigit`.

Now, tag prefixes prepend the default filter with `TAG_PREFIX[-v]*`.
This makes the filterer search for the defined tag prefix while
optionally accounting for any extra dashes or v's (for legacy branches
that use `v1.0` or something). So, the contract is now that a tag must
always include a digit for it to be included in versioning.