Fix configuration cache issues
This one's a bit weird. Gradle stores *the effects* of providers in
order to execute them without needing to configure the project. However,
in order to get the execution path for the Java launcher, we need access
to the JavaToolchainService, which is usually injected at configuration.
The problem is that this service is not available if we are skipping
configuration (i.e. configuration cache).

The solution is to not worry about setting a Java launcher executable
path. This plugin requires Gradle 9 which itself requires Java 17. Git
Version is built using Java 17. We can count on the daemon JVM being at
least version 17 and don't need to worry about setting the Java launcher
in the JavaExec spec ourselves.