Split off GitVersion and Changelog plugins into two separate jars
The Changelog plugin now exists in
`:gitversion-gradle:changelog-gradle`. This was done because the
GitVersion plugin can be applied onto any type (including settings)
while Changelog can only be applied onto projects. If a consumer were to
apply the GV plugin in `settings.gradle` and then later apply the
Changelog plugin in `build.gradle`, it could cause an error where the
plugin could already be on the classpath with an unknown version.

The only code shared by these two projects is a single class
`GitVersionTools` that has the definition for the tool to download and
invoke. This has been split off into `:gitversion-gradle:common` and is
shadowed by both projects.