Optimize usage of Git Version plugin in settings.gradle
If `net.minecraftforge.gitversion` is applied in
`settings.gradle(.kts)`, the `gitversion` extension is added to both the
`Settings` and `Gradle` objects. Subprojects can access
`gradle.gitversion` (or
`gradle.extensions.findByType(GitVersionExtension)` with an import
statement) to access the root project's version for free.

Additionally, all projects in the build will inherit the version
(`gitversion.tagOffset) found by Git Version (see
`Gradle#beforeProject`).

Projects that apply the plugin will first check if
`project.layout.projectDirectory` is the same as
`buildLayout.rootDirectory`. If it is and the Gradle instance has a
`gitversion` extension, it is directly added to the project without
recreating it.

This avoids the need to call Git Version multiple times just to get the
same version. Extremely useful for subprojects that do not need to
calculate their own version number.