Added some gradle properties
Properties used by Forge's Gradle plugins can be set either in the
`gradle.properties` file or as a system property, with precedence given
to the former. Both are checked, mostly to aid the usage of CI systems.
All properties are disabled by default

`net.minecraftforge.gitversion.log.version`
- Enabled if set to `true`.
- Logs the project's version after evaluation.
  - This uses the project's logger, so Gradle knows to keep the logged
version under the "Configure project :(project)" for the project it's
writing it for.